+++ src/EnvMapping/Lando.php
@@ -1,4 +1,4 @@
-<?php declare(strict_types=1);
+<?php
 
 namespace Druidfi\Omen\EnvMapping;
 
@@ -7,7 +7,8 @@ class Lando extends EnvMappingAbstract
   /**
    * @see https://github.com/lando/lando/blob/main/docs/config/env.md
    */
-  public function getEnvs() : array {
+  public function getEnvs() : array
+  {
     $lando_info = $this->getLandoInfo();
     $lando_host = $this->getLandoHost();
 
@@ -23,11 +24,13 @@ class Lando extends EnvMappingAbstract
     ];
   }
 
-  private function getLandoHost() : string {
+  private function getLandoHost() : string
+  {
     return getenv('LANDO_APP_NAME') . '.' . getenv('LANDO_DOMAIN');
   }
 
-  private function getLandoInfo() : array {
+  private function getLandoInfo() : array
+  {
     return json_decode(getenv('LANDO_INFO'), TRUE);
   }
 }
