+++ src/EnvMapping/Wodby.php
@@ -1,4 +1,4 @@
-<?php declare(strict_types=1);
+<?php
 
 namespace Druidfi\Omen\EnvMapping;
 
@@ -9,7 +9,8 @@ class Wodby extends EnvMappingAbstract
   /**
    * @see https://wodby.com/docs/infrastructure/env-vars/
    */
-  public function getEnvs() : array {
+  public function getEnvs() : array
+  {
     return [
       'APP_ENV' => $this->getAppEnv(),
       'DRUPAL_DB_DRIVER' => getenv('DB_DRIVER'),
